+++ common/EarlyReturnWithCode.cc
@@ -1,0 +1,10 @@
+#include "EarlyReturnWithCode.h"
+
+using namespace std;
+
+namespace sorbet {
+
+EarlyReturnWithCode::EarlyReturnWithCode(int returnCode)
+    : SorbetException("early return with code " + to_string(returnCode)), returnCode(returnCode){};
+
+}
