+++ main/lsp/protocol.cc
@@ -1,5 +1,6 @@
 #include "absl/synchronization/mutex.h"
 #include "absl/synchronization/notification.h"
+#include "common/EarlyReturnWithCode.h"
 #include "common/Timer.h"
 #include "common/web_tracer_framework/tracing.h"
 #include "core/lsp/TypecheckEpochManager.h"
@@ -11,7 +12,6 @@
 #include "main/lsp/json_types.h"
 #include "main/lsp/lsp.h"
 #include "main/lsp/watchman/WatchmanProcess.h"
-#include "main/options/options.h" // For EarlyReturnWithCode.
 
 using namespace std;
 namespace spd = spdlog;
@@ -126,7 +126,7 @@ optional<unique_ptr<core::GlobalState>> LSPLoop::runLSP(shared_ptr<LSPInput> inp
             auto params = make_unique<ShowMessageParams>(MessageType::Error, msg);
             config->output->write(make_unique<LSPMessage>(
                 make_unique<NotificationMessage>("2.0", LSPMethod::WindowShowMessage, move(params))));
-            throw options::EarlyReturnWithCode(1);
+            throw EarlyReturnWithCode(1);
         }
 
         // The lambda below intentionally does not capture `this`.
@@ -220,7 +220,7 @@ optional<unique_ptr<core::GlobalState>> LSPLoop::runLSP(shared_ptr<LSPInput> inp
                     if (taskQueue->errorCode != 0) {
                         // Abnormal termination. Exit immediately.
                         typecheckerCoord.shutdown();
-                        throw options::EarlyReturnWithCode(taskQueue->errorCode);
+                        throw EarlyReturnWithCode(taskQueue->errorCode);
                     } else if (taskQueue->pendingTasks.empty()) {
                         // Normal termination. Wait until all pending requests finish.
                         break;
