+++ main/main.cc
@@ -4,7 +4,7 @@
 int main(int argc, char *argv[]) {
     try {
         return sorbet::realmain::realmain(argc, argv);
-    } catch (sorbet::realmain::options::EarlyReturnWithCode &c) {
+    } catch (sorbet::EarlyReturnWithCode &c) {
         return c.returnCode;
     } catch (sorbet::SorbetException &e) {
         return 1;
