+++ main/options/options.cc
@@ -953,7 +953,4 @@ void readOptions(Options &opts,
     }
 }
 
-EarlyReturnWithCode::EarlyReturnWithCode(int returnCode)
-    : SorbetException("early return with code " + to_string(returnCode)), returnCode(returnCode){};
-
 } // namespace sorbet::realmain::options
